0709 11:50자연 인풋매니저에 상호작용키 추가(스테이션전용) savedata enum통일 안된거  통일화

+++ Assets/2.Scripts/Combat/Skill/SkillSystem.cs
@@ -192,7 +192,7 @@ public class SkillSystem : MonoBehaviour
 			Skill skill = _ownedSkills[i];
 			result[i] = new SavedSkill
 			{
-				skillId   = (int)skill.SkillId,
+				skillId   = skill.SkillId,
 				slotIndex = FindSlotIndexOf(skill as ActiveSkill)
 			};
 		}
